refactor ProductVariantsTable
