+++ project-base/storefront/public/locales/cs/common.json
@@ -174,7 +174,6 @@
     "Monday": "Pondělí",
     "My account": "Můj účet",
     "My orders": "Mé objednávky",
-    "Name": "Název",
     "name ascending": "Název A-Z",
     "name descending": "Název Z-A",
     "Need advice?": "Potřebujete poradit?",
