+++ project-base/storefront/public/locales/en/common.json
@@ -170,7 +170,6 @@
     "Monday": "Monday",
     "My account": "My account",
     "My orders": "My orders",
-    "Name": "Name",
     "name ascending": "Name A-Z",
     "name descending": "Name Z-A",
     "Need advice?": "Need advice?",
