+++ project-base/storefront/public/locales/sk/common.json
@@ -174,7 +174,6 @@
     "Monday": "Pondelok",
     "My account": "Môj účet",
     "My orders": "Moje objednávky",
-    "Name": "Názov",
     "name ascending": "Názov A-Z",
     "name descending": "Názov Z-A",
     "Need advice?": "Potrebujete poradiť?",
